Require semicolons in WIT files (#18)
This is a step further in the plan to add semicolons to WIT files everywhere. Now that tooling should have support for semicolons this updates the github action here to, by default, require semicolons. That way when proposals update to a new version of this action they'll by default require semicolons as well.