Add binary artifacts to WASI repo's CI
This commit updates the action configuration to produce a `*.wasm`
binary artifact to be published on GitHub Releases on tags for the time
being. This artifact is additionally uploaded for all CI runs if
necessary as well. At the same time this updates to use
`bytecodealliance/actions` to install `wit-bindgen` and `wasm-tools` to
avoid the need for `cargo install`.